Created SimpleMotorTest command
Created SimpleMotorTest command. This command is scheduled when the "A" button on an Xbox controller is pressed. It moves two motors (in the intake subsystem) depending on the reading from the color sensor. Command ends when color sensor detects red. Command intended to test both Talon and Spark motor controllers. NOTE: robotmap needs to be updated to correct values.
Update to Contants and Drivetrain: removed 2 motors from drivetrain